idpf: Add Inter-Device Communication interface file
Added new file 'idpf_eth_idc.h'.

Added idpf_eth_shared as common information to share between main
and ethernet device. Moved a member from idpf_adapter in new
eth shared structure, and moved some other structs: idpf_flags,
idpf_vector_info, idpf_intr_reg, idpf_q_vector into new file
'idpf_eth_idc.h' to lay foundation work for inter device
communication (idc) between main and ethernet device.

There is no change in the functionality before and after the change.
